face_api_demo/requirements.txt
flask==3.0.0
flask-cors==4.0.0
deepface==0.0.79
werkzeug==3.0.1
supabase==2.3.0
python-dotenv==1.0.0
